memcache-monitor/requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
PyYAML==3.10
Werkzeug==0.9.4
backports.ssl-match-hostname==3.4.0.2
ipdb==0.8
ipython==1.2.1
itsdangerous==0.23
matplotlib==1.3.1
nose==1.3.1
numpy==1.8.0
pyparsing==2.0.1
python-dateutil==2.2
six==1.6.1
tornado==3.2
wsgiref==0.1.2
